requirements.txt
astroid==2.2.4
autopep8==1.4.3
Django==2.1.7
django-cors-headers==2.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
Faker==1.0.2
isort==4.3.12
lazy-object-proxy==1.3.1
mccabe==0.6.1
nanoid==2.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
text-unidecode==1.2
typed-ast==1.3.1
wrapt==1.11.1
